chore: job detail
